Rename auxiliary to tables
