+++ CMakeLists.txt
@@ -6,11 +6,11 @@ set(CMAKE_C_STANDARD 90)
 include_directories(include)
 
 add_executable(Assembler_SemiCompiler
-        include/auxiliary.h
+        include/tables.h
         include/constants_tables.h
         include/first_pass.h
         src/assembler.c
-        src/auxiliary_functions.c
+        src/tables.c
         src/first_pass.c
         src/output.c
         src/second_pass.c
