+++ src/first_pass.c
@@ -1,6 +1,6 @@
 #include "../include/utils.h"
 #include "../include/validators.h"
-#include "../include/auxiliary.h"
+#include "../include/tables.h"
 #include "../include/first_pass.h"
 #include "../include/constants_tables.h"
 #include "../include/globals.h"
