+++ src/validators.c
@@ -3,7 +3,7 @@
 */
 
 #include "../include/validators.h"
-#include "../include/auxiliary.h"
+#include "../include/tables.h"
 #include "../include/status_codes.h"
 
 int validate_input(int argc, char** argv)
